move gribbles code to separate dir
