+++ src/gribbles/WordResult.js
